rend

+++ src/server/routes/agentRouter.ts
@@ -815,8 +815,8 @@ export function createAgentRouter(agentPool: Pool, scheduler?: AgentScheduler, b
            ROUND(AVG(hit_at_rank)::numeric, 1)::float          AS avg_rank,
            ROUND(AVG(optimal_n::float / 100), 4)::float        AS baseline,
            ROUND(
-             AVG(CASE WHEN hit THEN 1.0 ELSE 0.0 END) -
-             AVG(optimal_n::float / 100), 4
+             (AVG(CASE WHEN hit THEN 1.0 ELSE 0.0 END) -
+              AVG(optimal_n::float / 100))::numeric, 4
            )::float                                             AS vs_azar
          FROM hitdash.pair_recommendations
          WHERE game_type = $1
